fix potential underflow bug

+++ src/burnchains/burnchain.rs
@@ -723,8 +723,8 @@ impl Burnchain {
             return Ok(burn_chain_tip);
         }
 
-        let total = sync_height - self.first_block_height;
-        let progress = (end_block - self.first_block_height) as f32 / total as f32 * 100.;
+        let total = sync_height as f32 - self.first_block_height as f32;
+        let progress = (end_block as f32 - self.first_block_height as f32) / total as f32 * 100.;
         info!(
             "Syncing STACKS MAINCHAIN blocks: {:.1}% ({} to {} out of {})",
             progress, start_block, end_block, sync_height
